Installation instructions by Goose/xAI
These are the instructions to install the plugin with Lazy.nvim. All you
have to do is tell Lazy.nvim where to find the plugin on GitHub and it
will download the project.

The "this is where to get the plugin" instruction file goes in
~/.config/nvim/lua/plugins and then Lazy.nvim will download them to
~/.local/share/nvim/lazy/

These instructions were generated by Goose using xAI.